refactor :: auth 컴포넌트 간소화
