+++ src/components/auth/loginForm/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as S from './style'
+import * as S from '../style'
 import TextInput from "src/components/textInput";
 import {useAuth} from "src/hooks/useAuth";
 import {useNavigate} from "react-router-dom";
@@ -8,7 +8,7 @@ const LoginForm = () => {
     const {...auth} = useAuth()
     const navigate = useNavigate();
     const goBack = () => navigate(-1);
-    const goForgotPassword = () => navigate(-1);
+    const goForgotPassword = () => navigate('/forgot-PW');
     const goSignup = () => navigate('/signup')
     return (
         <S.Wrapper>
@@ -30,14 +30,14 @@ const LoginForm = () => {
                 value={auth.user.password}
                 change={auth.changeInput}
                 size={30}
-                warn={auth.isPassword ? null : `공백을 제외한 영어 대/소문자, 숫자를 4~30자 사이로 입력하여야 합니다` }
+                warn={auth.isPassword ? null :'공백을 제외한 영어 대/소문자, 숫자를 4~30자 사이로 입력하여야 합니다' }
                 isPassword
             />
 
             <S.ButtonContainer>
-                <S.FormButton onClick={auth.signIn}>
+                <S.FormButton2 onClick={auth.signIn}>
                     <S.ButtonText>로그인</S.ButtonText>
-                </S.FormButton>
+                </S.FormButton2>
                 <S.Text onClick={goForgotPassword}>비밀번호 찾기</S.Text>
                 <S.Text className={'color'} onClick={goSignup}>회원가입</S.Text>
             </S.ButtonContainer>
